fix: context actions types
