+++ packages/web-pkg/src/components/ContextActions/index.ts
@@ -2,3 +2,4 @@ export { default as ActionMenuItem } from './ActionMenuItem.vue'
 export { default as ContextActionMenu } from './ContextActionMenu.vue'
 export { default as ContextMenuQuickAction } from './ContextMenuQuickAction.vue'
 export { default as ActionMenuDropItem } from './ActionMenuDropItem.vue'
+export * from './types'
